Refactor and add comments to plugin
